project: cleanup.
remove unused imports and comments, fix typos etc,
rename files and reorder in project.
