+++ app/src/main/java/com/huji/foodtricks/buddies/EventsTabsActivity.java
@@ -1,41 +1,18 @@
 package com.huji.foodtricks.buddies;
 
-import android.app.Activity;
-import android.content.Context;
 import android.content.Intent;
 import android.support.design.widget.TabLayout;
 import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 
-import android.widget.TextView;
-
-import java.util.ArrayList;
-
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.view.Menu;
 
 
-public class EventsTabs extends AppCompatActivity implements TabLayout.OnTabSelectedListener,ViewPager.OnPageChangeListener {
+public class EventsTabsActivity extends AppCompatActivity implements TabLayout.OnTabSelectedListener,ViewPager.OnPageChangeListener {
 
     ViewPager vp;
     TabLayout tabLayout;
@@ -56,12 +33,10 @@ public class EventsTabs extends AppCompatActivity implements TabLayout.OnTabSele
             }
         });
 
-        //VIEWPAGER
-        vp= (ViewPager) findViewById(R.id.mViewpager_ID);
+        vp = (ViewPager) findViewById(R.id.mViewpager_ID);
         this.addPages();
 
-        //TABLAYOUT
-        tabLayout= (TabLayout) findViewById(R.id.mTab_ID);
+        tabLayout = (TabLayout) findViewById(R.id.mTab_ID);
         tabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
         tabLayout.setupWithViewPager(vp);
         tabLayout.setOnTabSelectedListener(this);
@@ -71,11 +46,10 @@ public class EventsTabs extends AppCompatActivity implements TabLayout.OnTabSele
     private void addPages()
     {
         ViewPagerAdapter pagerAdapter=new ViewPagerAdapter(this.getSupportFragmentManager());
-        pagerAdapter.addFragment(new FutureEvents());
-        pagerAdapter.addFragment(new PastEvents());
-        pagerAdapter.addFragment(new PendingEvents());
+        pagerAdapter.addFragment(new FutureEventsTabFragment());
+        pagerAdapter.addFragment(new PastEventsTabFragment());
+        pagerAdapter.addFragment(new PendingEventsTabFragment());
 
-        //SET ADAPTER TO VP
         vp.setAdapter(pagerAdapter);
     }
