+++ app/src/main/java/com/huji/foodtricks/buddies/Models/EventModel.java
@@ -1,4 +1,6 @@
-package com.huji.foodtricks.buddies;
+package com.huji.foodtricks.buddies.Models;
+
+import com.huji.foodtricks.buddies.EventAttendanceProvider;
 
 import java.io.Serializable;
 import java.util.ArrayList;
