+++ app/src/main/res/layout/activity_create_event.xml
@@ -131,7 +131,7 @@
         android:layout_marginEnd="8dp"
         android:layout_marginBottom="8dp"
         android:clickable="true"
-        android:onClick="fabClicked"
+        android:onClick="updateDatabaseWithNewEvent"
         app:backgroundTint="@android:color/darker_gray"
         app:backgroundTintMode="multiply"
         app:layout_constraintBottom_toBottomOf="parent"
