+++ app/src/main/res/layout/activity_events_list.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".EventsTabs">
+    tools:context=".EventsTabsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
