+++ app/src/main/res/layout/fragment_events_tabs.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".EventsTabs">
+    tools:context=".EventsTabsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
