+++ app/src/main/res/values/strings.xml
@@ -68,7 +68,7 @@
     <string name="app_description">Hanging out with friends was never easier</string>
     <string name="friends_hanging">Friends Hanging</string>
     <string name="where_title">where?</string>
-    <string name="tantative_msg">Tantative</string>
+    <string name="tentative_msg">Tentative</string>
     <string name="decline_msg">decline</string>
     <string name="approve_msg">approve</string>
     <string name="tab_text_1">tab_text_1</string>
